+++ R/prep_sosdata.R
@@ -21,8 +21,14 @@
 #' @param remove Should the individual DIA (except HDIA), OP, ekod variables
 #'   be removed after constructing combined variable (to save space)?
 #'   Default is TRUE.
-#' @param impute To be implemented. kom ihåg att kolla om finns värden i INDATUMA
+#' @param impute If patreg = "dors" imputes missing month and date in DODSDAT.
+#'   Possibly to be implemented in patreg. kom ihåg att kolla om finns värden i INDATUMA
 #'   (text) men inte i INDATUM (numerisk). i så fall imputera.
+#' @param diavar The starting characters of variables containing diagnosis (used if registry = patreg).
+#' @param opvar The starting characters of variables containing operations (used if registry = patreg).
+#' @param evar The starting characters of variables containing ekod (used if registry = patreg).
+#' @param orsakvar The starting characters of variables containing causes of death (used if registry = dors).
+#' @param mainvar The main variable that should be saved in the data even if remove = TRUE.
 #'
 #' @seealso \code{\link{create_sosvar}}
 #'
@@ -38,43 +44,57 @@
 #'   dagkir %>% mutate(source = "DK")
 #' )
 #' }
-#' sos_data <- prep_sosdata(sos_data)
+#' sos_data <- prep_sosdata(sos_data, evar = "ekod")
+#' dors_data <- prep_sosdata(dors_data, registry = "dors", impute = FALSE)
 #' @export
 
 prep_sosdata <- function(sosdata,
                          registry = "patreg",
                          utdatum = FALSE,
                          remove = TRUE,
-                         impute = TRUE) {
-  if (!registry %in% c("patreg", "dors")) stop("registry is not patreg or dors")
+                         impute = TRUE,
+                         diavar = c("HDIA", "DIA"),
+                         opvar = "OP",
+                         evar = "EKOD",
+                         orsakvar = c("ULORSAK", "MORSAK"),
+                         mainvar = ifelse(registry == "patreg", "HDIA", "ULORSAK")) {
+  if (!registry %in% c("patreg", "dors")) stop("registry must be either patreg or dors")
 
   if (registry == "patreg") {
     sosdata <- sosdata %>%
       mutate_at(
-        vars(dplyr::matches("DIA|OP|ekod")),
+        vars(dplyr::starts_with(c(diavar, opvar, evar))),
         list(~ tidyr::replace_na(., ""))
-      ) %>%
-      # create temp HDIA variable (not containing DIA) so is not removed in unite step
-      mutate(tmp_H = paste0(" ", HDIA)) %>%
-      tidyr::unite(DIA_all, contains("DIA", ignore.case = FALSE),
+      )
+    if (remove) {
+      sosdata <- sosdata %>%
+        # create temp HDIA variable (not containing DIA) so is not removed in unite step
+        mutate(tmp_mainvar = paste0(" ", !!sym(mainvar)))
+    }
+    sosdata <- sosdata %>%
+      tidyr::unite(DIA_all, starts_with(diavar, ignore.case = FALSE),
         sep = " ",
         remove = remove
       ) %>%
-      tidyr::unite(OP_all, contains("OP", ignore.case = FALSE),
+      tidyr::unite(OP_all, starts_with(opvar, ignore.case = FALSE),
         sep = " ",
         remove = remove
       ) %>%
-      tidyr::unite(ekod_all, contains("ekod", ignore.case = FALSE),
+      tidyr::unite(ekod_all, starts_with(evar, ignore.case = FALSE),
         sep = " ",
         remove = remove
       ) %>%
-      # rename HDIA back to org name
-      rename(HDIA = tmp_H) %>%
       mutate(
         DIA_all = paste0(" ", stringr::str_squish(DIA_all)),
         OP_all = paste0(" ", stringr::str_squish(OP_all)),
         ekod_all = paste0(" ", stringr::str_squish(ekod_all))
       )
+    if (remove) {
+      sosdata <- sosdata %>%
+        # rename HDIA back to org name
+        rename(!!mainvar := tmp_mainvar)
+    }
+
     if (utdatum) {
       sosdata <- sosdata %>%
         mutate(
@@ -96,23 +116,37 @@ prep_sosdata <- function(sosdata,
 
   if (registry == "dors") {
     sosdata <- sosdata %>%
+      filter(!is.na(DODSDAT)) %>%
       mutate_at(
-        vars(dplyr::matches("ORSAK")),
+        vars(dplyr::starts_with(orsakvar)),
         list(~ tidyr::replace_na(., ""))
-      ) %>%
-      # create temp ULORSAK variable (not containing ORSAK) so is not removed in unite step
-      mutate(tmp_U = paste0(" ", ULORSAK)) %>%
-      tidyr::unite(ORSAK_all, contains("ORSAK", ignore.case = FALSE),
+      )
+    if (remove) {
+      sosdata <- sosdata %>%
+        # create temp ULORSAK variable (not containing ORSAK) so is not removed in unite step
+        mutate(tmp_mainvar = paste0(" ", !!sym(mainvar)))
+    }
+    sosdata <- sosdata %>%
+      tidyr::unite(ORSAK_all, starts_with(orsakvar, ignore.case = FALSE),
         sep = " ",
         remove = remove
       ) %>%
-      # rename ULORSAK back to org name
-      rename(ULORSAK = tmp_U) %>%
       mutate(
         ORSAK_all = paste0(" ", stringr::str_squish(ORSAK_all))
       )
-    sosdata <- sosdata %>%
-      filter(!is.na(DODSDAT))
+    if (remove) {
+      sosdata <- sosdata %>%
+        # rename ULORSAK back to org name
+        rename(!!mainvar := tmp_mainvar)
+    }
+    if (impute) {
+      sosdata <- sosdata %>%
+        mutate(sos_deathdtm = lubridate::ymd(case_when(
+          substr(DODSDAT, 5, 8) == "0000" ~ paste0(substr(DODSDAT, 1, 4), "0701"),
+          substr(DODSDAT, 7, 8) == "00" ~ paste0(substr(DODSDAT, 1, 6), "15"),
+          TRUE ~ DODSDAT
+        )))
+    }
   }
   return(sosdata)
 }